move public API into include directory
