+++ core/common/Exceptions.cpp
