+++ core/common/tests/ExceptionsTest.cpp
